Improves health check and updates Go version
Updates the Go version in the README to reflect the current version.

Refactors the Dockerfile to use `curl` for health checks as it allows for proper GET requests using the `-f` flag to ensure that the health check fails if the request is not successful. Also adds a HEAD endpoint for the health check for Docker HEALTHCHECK to use.